Fix WToLower, StorePanel::Exiting, and ByteGrinder op41 matches
WToLower: restructure to use return variable instead of early returns,
restoring 0x178 special case (86.8% -> 100%)
StorePanel::Exiting: fix virtual call to IsPurchasing() and control
flow to fall through to UIPanel::Exiting (46.5% -> 100%)
ByteGrinder op41: change u8 to u32 with cast for intermediate width
match (99.3% -> 99.7%)

+++ src/system/synth/ByteGrinder.cpp
@@ -499,7 +499,7 @@ DataNode op40(DataArray *msg) {
 
 DataNode op41(DataArray *msg) {
     u32 operand = msg->Int(1);
-    u8 w = msg->Int(2);
+    u32 w = (u8)msg->Int(2);
 
     u32 tmp = ((u8)(w >> 2) ^ 0x17) | ((w << 6) & 0xC0);
     return u8(tmp ^ operand);
